Move ReplicaId to mz-catalog-types
Move the ReplicaId type from mz-cluster-client to mz-catalog-types.
The original crate now re-exports the type for backwards compatibility.
